Use the default directory C:\Run for GetFileRequest when source file has no path Added GetDeviceFile in SVRemoteControl which was missing Remove warnings in SVRC build Issue SVO-1978

+++ SVRemoteControl/SVControlCommands.cpp
@@ -86,7 +86,7 @@ SVControlCommands::SVControlCommands(NotifyFunctor p_Func)
 	if (bModul)
 	{
 		IniFile.append(szPath);
-		int pos = IniFile.find_last_of('\\');
+		size_t pos = IniFile.find_last_of('\\');
 		if (pos != std::string::npos)
 		{
 			IniFile.erase(pos + 1);
